requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
tqdm
pillow
jupyter